Tree resizing, elision, and tooltips
Improve tree section resizing, allow manual resizing in some cases, and add tooltips for elided text.
In Filesystems, the mount path column can be indefinitely wide, so move it to the end and set to stretch.
Use special logic to resize the first column where it can be wide: set that section to the minimum of the contents width or the available width in the window.  Use that method for FilesystemsWindow (the "available" width is shared with the last, stretched, mount path section) and UnreadableDirsWindow.  This gives the best chance of the first column not being elided without taking up unnecessary whitespace.  Let the columns be resized further by the user.
Display tooltips for elided column text in some cases: for the first and last columns of FilesystemsWindow, the first column of UnreadableDirsWindow, and the last column of LocateFilesWindow and LocateFileTypeWindow, override QTreeWidgetItem::data to calculate the section width, indentation, icon, and text sizes; for the first column of the main DirTreeView, override viewportEvent to compare the sizeHint and visualRect widths; and for the first (only) column of the OpenDirDialog directory tree, use a delegate to override helpEvent and compare the sizeHint and visualRect widths.
Put the common functions in FormatUtil; HeaderTweaker knows about headers but not really the trees themselves.
FileAgeStats, FileTypeStats, and the two FileSizeStats tables remain completely sized to contents with no tooltips; all columns are relatively small and relatively consistent sizes so they will normally all fit in the available space.