Elision and tooltips in FileDetailsView
Elide the tooltip target to the configurable label limit even if it isn't an absolute path, and re-elide even if it is a shortened absolute path, for people that really don't like scrollbars.
Also limit the package label to the same limit.  All other fields either have relatively fixed short lengths or are under the control of the user (eg. category names).
Add tooltips for all elided labels.
Change some instances of int to FileCount in FileDetailsView.
Rename two functions that were overloads, both taking integers (FileCount and FileSize) that may or may not be ambiguous depending on the platform and future development.