Cleanup confirmation dialog elision
Improve the elision and sizing of the confirmation dialog for Cleanups.
Instead of eliding and padding the title to a fixed maximum width, calculate the maximum width using the same algorithm as Qt.
Work in pixels and only convert to spaces when actually padding the title line.  Be a bit less sloppy and take into account the item type prefix and HTML entities.