Improve Breadcrumb elision
The Breadcrumb component elision was pretty brutal, slashing the length of one component at a time to the absolute minimum regardless of how much space saving was needed.  The minimum and maximum sizes were also fixed.
Calculate the maximum length based on the available space.
Elide the longest component by half or the necessary space to fit, and keep doing that until the breadcrumbs will fit in the screen.  Components will be elided more evenly and only by about as much as necessary to fit.
Only stop eliding at one character long, the ellipsis.
Re-calculate whenever the screen sizing changes.