Path elision in FileDetailsView
The character-length label limit is not very good at preventing a horizontal scrollbar in the file details panel, and is fairly pointless if it still produces a scrollbar.
Replace this limit with a simple boolean elide-or-not that elides the path labels to fit within the panel width.  There can still be a scrollbar if other fields are too narrow, but they are fairly strictly limited and in any case numbers and dates are not well-suited to being elided.  The user can choose a panel reliably wide enough for them.  Or not.
Respond to any size changes by re-calculating the available panel width and filling it in again from scratch.
Move all the gruntwork of filling in labels into the unnamed namespace, leaving just the public API and a few key protected functions.