Tighten up elision in the dialog headers
Or loosen up.  The code used a hard-coded "few extra pixels" to account for margins, frames, etc, but in some cases it wasn't enough.  Layouts and margins can vary depending on themes, dpi settings, etc.
So calculate the exact available space for the label text.  Use the "last pixel" algorithm so the label.x() can account for spacing, other widgets, on the left.  Usually the last pixel is the dialog width minus the righthand layout margin, but the find files dialog is more complex with nested layouts and a frame (hence an additional indentation of half the 'x' character).  Still add a few pixels to let the user shrink the dialog; the exact amount (plus 1 pixel) doesn't seem to be enough for some trailing characters, possibly due to bearing.
Catch font changes for the breadcrumbs which might change the elision.
Catch file details font and palette changes in the panel instead of the main window.
Add an overload in FormatUtils for the horizontal advance of a single character.