Right-elide the name column in DirTreeView
In common with every other tree view column, elide right in the main directory tree.  This is easier to spot and any difficulties with lacking the end of a path are eased by having the tooltip.
Labels not in trees currently remain elided in the middle: I think that's definitely best for the breadcrumbs, maybe not for file details and dialog titles.