Don't add DirInfo from fstatat error to dot entry
An "empty" DirInfo is created when fstatat() fails.  For fairly obvious reasons, the mode does not actually match to a directory, so DirInfo adds the child to the dot entry.  Make DirInfo check whether the child is a DirInfo instead, so the failed directory doesn't appear under <Files> in the tree.