Better handling of unread items in cache files
When an item has a read (fstatat) error, a DirInfo object is created which may or may not represent the actual filesystem object type.  The mode is likely to be 0.  When a cache file containing such an object is read, it will fail to be recognised as a DirInfo.  The resulting FileInfo object will not be understood by the rest of the program.
Add an "unread" case for read state DirError.  Force all items with "unread" entries to create DirInfo objects regardless of the type or mode.  And obviously mark unread/readerror as DirError.