Control characters, html tags, etc
Control characters in labels, most obviously carriage return and line feed, can wreak havoc with the layout.  They can also blow out the line height in lists if they are in the first row (for trees) or the last (for list views).  Character sequences that look like html tags can also cause labels and tooltips to render as rich text, and the sequences are interpreted as actual html, which again can wreak havoc.  At the other extreme, sequences of whitespace in rich text are, by default, collapsed to a single space, which may not be the best way to represent, for example, a pathname with a sequence of multiple spaces.
Use QRegularExpressionValidator to prevent users entering control characters in mime category names, cleanup titles, exclude rule patterns, and find file search strings.
Force labels which may show values containing arbitrary characters, including Linux pathnames and categories, to render as plain text, using the textFormat.  Literal carriage returns and line feeds are converted to spaces for labels that must appear in a single line.  All the trees have uniform heights so, assuming the first item is on a single line, items with line feeds will be vertically elided, that is only the text up to the line feed will appear, followed by an ellipsis.  Remove carriage returns and line feeds from paths in all trees except DirTreeView.  For DirTreeView, make a special case by removing carriage returns and line feeds from the first (ie. root) row as the simplest way to avoid a very rare but catastrophic situation.
Provide tooltips for tree rows that are elided vertically or have had carriage returns and line feeds converted to spaces.  Leave carriage returns and linefeeds in the tooltip path so that they may appear on multiple lines.  Provide the same sort of tooltip for paths in the file details panel, but not for path labels elsewhere.  Prevent these tooltips being interpreted as rich text which would cause them to wrap differently and potentially render embedded html, and allow them to wrap even without word breaks, by inserting special formatting characters into the tooltip text.  Expand tooltips for DirTreeView to any elided column, provided that the model doesn't offer a tooltip.
In BreadcrumbNavigator, remove carriage returns and line feeds to stop the label breaking onto multiple lines. Wrap it to stop whitespace collapsing to a single space character.  Escape the individual crumb names to prevent embedded tags being treated as html.  Fix a problem with trying to elide a path before the central widget has been laid out, as happens when the root path has no sub-directories and the event loop never gets to spin.