section() gets the indexOf treatment
QString::section() is flexible, simple, and powerful, but for many cases it is inefficient overkill: it always splits the entire string to create a string list.  So for most cases, replace it with last/indexOf and mid/left/right.
Leave the snap package extraction in MountPoints, never very many and it is a non-trivial substring.
Use the existing SysUtil::baseName in PkgManager, speeds up package reading by about 5% when IO is fully cached.
Replace section( ' ', 0, 0 ) in Cleanup, a trivial case, although not going to be run thousands of times  Also, just run it once per app instead of once for every path in the environment.
Replace the suffix extraction in MimeCategorizer.  Also refactor the loop so it doesn't even do the substring extraction unless there is a suffix.  This is a central part of the categorizer and performance is apparently improved by about 20% in FileTypeStats, somewhat less overall in Treemap.
Replace the suffix extraction in FileTypeStats.  The performance gain is considerable, but only for non-category files, so overall just about a 10% gain depending on the workload.