Handle zero-sized pies in HistogramView
Normally, zero-sized pies are only created for empty sets and no pies appear if percentiles are cut off, which was acceptable.  But if there are files and all of them have zero bytes then one pie can appear for cutoff files but none for cutoff bytes.  If this happens when toggling the symlinks then there are resizing issues as the height of the cutoff panel changes unexpectedly.
So, make addPie accept both the slice and the pie being zero without causing divide by zero errors.  Now a pies appears as a complete blue circle if the slice and pie are both zero.