More string wrangling
Change a regular expression for extracting the directory portion of a path to indexOf.  This speeds up Cleanup execution by about 5x, very noticeable when processing thousands of files at once.
Relax the requirements for using DirTreeSuffixFilter so that the suffix only has to not contain wildcard characters (*, ?, and [).  This matches the definition used by MimeCategory, so extract the check into a single (inline) static function in Wildcard.
Extract some repeated string munching from inside loops in Logger and other minor tweaks to reduce copying of strings, none of them dramatic.