Fix file type searching
Replace non-suffix rules in FileTypeStatsWindow with the actual pattern used to assign that category.  There is still "<non-pattern rule>" for symlinks and executables identified by their file mode.
Extend searching in LocateFileTypeWindow and FileSizeStatsWindow to handle the non-suffix patterns as well as the executables and symlinks.  Uncategorised files with "<no extension">, including cruft, still cannot be searched.
Refine searching so that not just the pattern is considered, but also whether that filename is actually in that category.  So, for example a search for *.h in one category will no longer find ui_*.h files in a different category.
Much of this work revolves around using the WildcardCategory pair struct.