Cleanup after MimeCategory duplicate changes
Move WildcardCategory from MimeCategorizer into Wildcard now that it is used in other places.
Switch to "|" as the separator for the default category lists instead of comma ("/" might be even less likely to be in a pattern, but looks a bit odd).  Also remove whitespace between entries; whitespace is now significant in a pattern.
Use SignalBlocker to prevent premature duplicate checking when loading patterns for a new category.  This detects false positives, then "undetects" them so the user doesn't see anything, but this forces focus out of the list.