Remove the blocking of the asterisk key from DirTreeView
DirTreeView overrides keyPressEvent in order to catch the asterisk key that would otherwise open all descendents of the current item.  Ostensibly for performance reasons, but also because "we have better alternatives" with the expand to level actions.  But the asterisk hotkey operates in a quite different way; while the expand to level actions expand the whole tree to a particular level, asterisk only operates from the current item downwards.
So get rid of the key press event handler and allow astrisk to operate as expected for a QTreeView.
Also do some re-factoring, simplifications, etc.