Clean up SysUtil::runCommand and related functions
There is a runCommand overload that accepts a full command line string and attempts to parse it into a command and argument list.  The functions that use this overload already know the command and individual arguments, but join them together.  Stop doing this pointless make-work.  Add a PkgCommand structure for returning both a command string and args string list from the PkgManager sub-classes, drop the runCommand overload, and tweak tryRunCommand to do the same.