More robust BusyPopup event loop
In some situations, most obviously in Qt6 when starting a Pkg or Unpkg read from the command line, the simple processEvents() call fails to display the popup and there is an unpainted window until the packages have been read.
Instead, spin a local event loop (ie. QEventLoop::exec) which quits when a paint event is received for the label.