Throw final at everything
Mark all classes that are not inherited as final. Even the ones with no virtual functions.  There is no real downside in a standalone program like this.  There is some upside with the compiler getting to make some extra optimisations, and it flags up that those classes may not be suitable for inheriting.