Set sizePolicy for elided labels
The typical size policy of Preferred, or sometimes Maximum, for labels which can be elided causes the label to blow up to the screen width when the parent window is maximised and then normalised.  This happens when the size of the window is adjusted to account for the size of its children and is difficult to intercept in event handlers.  The relevant size hint is actually provided by the layout containing the label, so also cumbersome to override.
Setting the size policy to Ignored prevents the window from adjusting the size of the label layout to the width of the maximised window when it changes the window state.  It also improves resizing performance as the window is no longer pointlessly calculating the preferred size of the label.
One minor downside is that the breadcrumb navigator now expands to the available width and so a context menu on the blank area to the right of the label text now somewhat confusingly relates to the breadcrumb text instead of to the layout customisation.