Fix HisatogramView resizing
In the resize event handler, the viewport may not have been resized yet, so the rebuild is incorrect.  For manual window resizing, this is generally just a few pixels out, especially with the last resize event at the end of the resize, and this is easily hidden in the margins.  The options button show/hide resizes the viewport first, so that's OK.  But maximising and un-maximising, for those window managers that allow it on a dialog window, can be badly wrong.
Explicitly resize the viewport widget in the event handler before doing the rebuild.  The geometry is trivial, with no scrollbars and no margins, just a 1-pixel frame.
Also fix the P<sub>n</sub> string to actually use <sub> tags instead of relying on "vertical-align: sub" to also reduce the font size (which it does, but does not seem to be HTML standard and I don't know if it is even intentional for Qt.