Fix header resizing in LocateFilesWindow
LocateFilesWindow is unique in having its heading text passed in and depending on the type of locate being performed.  This includes "%1" to substitute the current search path, which may change if a fallback subtree path is used during a refresh.  The substitution, including fallback detection, was being performed for every resize which resulted in unexpected title changes if the original path no longer existed.
Store the original heading text in the class.  Substitute the current, possibly fallback, path, each time populate() is called, and store that in statusTip.  Re-elide the unchanging statusTip text for each resize (and other events).