DirReadJob permission checks
Use faccessat() instead of access() to test for directory read/execute permission.  This takes into account the effective uid and capabilities instead of simply whether the real uid is 0.  This fixes an uncommon issue where access() claims that a directory can be read, but opendir() fails.  Sych cases are now reported as permission denied instead of an unknown read error.
Also fix a compilation warning about _warnedAboutNtfsHardLinks being unused when VERBOSE_NTFS_HARD_LINKS is true, apparently because the compiler optimises away uses of the static variable that is set, but never read.  Take the opportunity to make the non-verbose log message more generic, stating that all NTFS hard link counts are being treated as 1 rather than just for the first file encountered.