Add a setting for whether to trust NTFS hard link counts
With the ntfs-3g driver, hard links counts may be reported incorrectly.  An option (posix_nlink) has been added to avoid this.  There is also a newer ntfs3 driver for mounting NTFS volumes in Linux.  So the problem is essentially fixed, but not everyone is using the newer driver yet and not everyone still using the older driver will want or be able to set the posix_nlink option.
So add a settings that flags whether NTFS hard links counts should be believed or reset to 1.  Make the default to trust them, but log a message if multiple hard links are found on an NTFS volume, whether trusted or not.