Rework the NTFS had link detection
Move the "one-off" check for NTFS into DirTree.  It is now done once per read, including refreshes, and simply logs a message that NTFS hard links are either being trusted or are being reset to 1.  No check in PkgReader.
Then in DirReadJob, if NTFS hard links are not configured as trusted, reset st_nlink to 1.  There is no NTFS checking in DirReadJob in the trusted case.  Continue to carry a cached value for whether the current job/directory is NTFS, but use a three-way NotChecked/NotNtfs/Ntfs flag instead of two bools.