Make package manager detection asynchronous
Detecting package managers, and in particular detecting a primary package manager, takes up most of the application startup time.  Although with fully-cached data this is only a few hundred milliseconds and not particularly noticeable, it can be tens of seconds with uncached data and a mediocre HDD.  And three package manager objects do this one after the other.
So make the external processes from the package managers start in the background.  Do this even before the main window is created, just to scavenge a little extra time.  The main window is not blocked and most functions, including reading directories, can be executed before the primary package manager is known.  The basic list of package managers is populated synchronously just using SysUtil::haveCommand, so finding an owning package can also be done before the primary package manager is known.
The packaged and unpackaged actions are disabled until the primary package manager processes complete.  As each process completes, it checks to see if the actions should be enabled.  If a package manager decides it is the primary, it sets the primary in PkgQuery and enables the actions.  The package read action is enabled even for non-primary package managers although PkgReader will block until all the check processes have finished so it can use a file list cache if possible.
Attempts to do a Pkg:/ or Unpkg:/ read from the command line will be blocked until the external check process finish or time out.  This is largely invisible behind the existing BusyPopup.
PkgQuery now carries the primary package manager explicitly instead of just as the first in the list.  It may be null.  A primary will only be set if it supports generating a file list cache and this doesn't need to be checked each time in PkgReader.
The timeout for these processes is increased to 30 seconds.  The previous 15 seconds was already borderline, and sometimes not long enough, and now three processes are being started in parallel, possibly at the same time as a filesystem read.  Owning package process timeouts are shorter and will be made configurable since no single value is really suitable for all cases.