Make unpackaged query more complete
Change the unpackaged query to ignore files from any package manager (only ones that support a file list cache for now).  Split PkgFileListCache into separate QMultiHash and QSet sub-classes.  Both are much simplified without all the flags.  PkgFileListCache.cpp is now redundant.

PkgManager always generates a QMultiHash.  Add a PkgQuery function to generate a PkgFileListCache from each package manager that can, and combine the filenames from those into a single QSet which is used as the filter.

The unpackaged action no longer requires a primary package manager, so it can be enabled immediately instead of when the check process completes.  Also enable the packaged query action, which can be done without a file list cache if necessary or if there are only a few packages.