Tweaking the tweaker
Make the HeaderTweaker context menu appear even when clicking on the part of the header to the right of the configured sections.  Obviously the first two actions are not shown in this case.
Make un-hidden sections appear in a consistent location.  Currently, this is the (right-hand) end.  Might be nice to remember where they were before they were hidden, but that opens a whole can of worms.  Might also be nice to remember the size details for each section per-layout.