Handle permission denied for item in readable directory
When stat fails for an individual item, a read error is set but this does not propagate up to its ancestors.  This can happen if permissions to the parent directory are OK, but permissions are denied on the item itself.  Since the mode is unknown, the item is not detected as a directory (and might not even be a directory) and _errSubDirs is not incremented.
So always increment _errSubDirs when there is a read error.
Also, inspect errno and set DirPermissionDenied for permissions failures and only use DirError for other more serious errors.