More package reading simplification
Drop PkgFileListCache::pkgManager.  CachePkgReadJob is now only created with a package for the package manager used for the cache (the primary).
Also drop the multi-arch and multi-version flags from PkgInfo.  There were (potentially) used for dpkg to generate a suitable query name, but are not actually needed.
Change QMultiMap in PkgReader::handleMultiPkg and DirTree::deleteSubtrees to QMultihash.  QMultiHash::uniqueKeys was only introduced in Qt 5.13, but prior to that there was QHash::uniqueKeys, so it is all good.  Ordered keys aren't needed and QMultiHash is a little more efficient.
Give some of the "major" dpkg commands a timeout of 30 seconds instead of 15 (or 5!) since they can take along time on the first run with uncached databases.  They may also be executed in parallel with other external processes such as RpmPkgManager::createFileListCache or the primary package check queries.  Keep the 5-second timeout for the owning package queries which should be working with cached data, perform relatively simple commands, and block the application (albeit with AdativeTimer).