Drop the "should be a directory"
The "should be a directory, but is not" warning is a holdover from when symlinks in the directory path were not resolved.  It was left in as a relatively harmless message that might pick up some unusual situation, but it gets triggered when there are directories in a package with no permissions.  This is a fairly common situation (I'm looking at you, policykit) that is not an error, so just get rid of that unhelpful message.  Stat failures are already reported, including missing items if the VerboseMissingPkgFiles setting is true, but permission denied is just ignored (there is a commented-out logDebug() message, if desired).