Avoid indenting the permissions in the details panel
The first character of the symbolic permissions string is a space for regular (non-sticky) files.  This causes the left-aligned text in the file details panel to appear indented.  The right-align DirTreeView is not affected.
Use a zero-width space instead, as the simplest solution, since the string is constructed from a series of individual characters.