Propagate all read errors up the tree
Read errors need to be propagated to ancestors by calling recalc(), which happens when they are marked dirty.  DirReadJob::handleStatError wasn't doing this and so the directory tree sometimes didn't indicate that there was a read error in one of its children.
Provide a DirInfo::setReadError function to both set the read state and mark as dirty so that the error counts are recalculated.  Call this from PkgReader (which was already setting dirty) and from DirReadJob.