Make PkgManager process timeouts more configurable
Provide an "owning package" timeout setting that is much faster than the 15-30 seconds of the main package management process commands; default five seconds.  Put this in PkgQuery so it can be accessed from any of the package managers.
Also move the rpm warning timeout setting to PkgQuery as a generic warning timeout, although it is only used in RpmPkgManager at the moment.
Also make the PkgQuery owning package cache size configurable: default 1,000.  The cache cost is always 1, so ignore it since that is the default for QCache::insert.