Zero is a valid directory size
"Normal" directories always take up space, but "virtual" ones such as those in /proc and /sys have a perfectly valid size of 0 bytes.  Display this in the file details panel instead of blindly hiding the entire own size row whenever a zero shows up.  It is already shown in the tree, but only noticeable in "empty" directories.  For directories in a package view, hide the row since the zero in those cases is an indication that we are ignoring the own size in that situation, not that the size actually is zero.

The uid/gid/perm fields are disabled when they are flagged as "unavailable".  This occurs in reads of old-format cache files and for some read errors.  Change so that those fields show as "available" (ie. enabled, not greyed-out) when there is a read error, but leave them empty since the contents are unknown.  This requires adding a new read state DirNoAccess for those cases where there is read permission on the parent, but not for the item itself (eg. the parent directory has permissions with no execute flag).  So for DirPermissionDenied, the own fields are known and shown, but for DirNoAccess they are unknown and blank.
For DirError and DirNoAccess, show the type field as "unknown" instead of blindly claiming it is a directory.  The uses of DirNoAccess are quite rare at the moment, but will become more common with additional changes to follow.

Change the display of aborted, or otherwise failed, packages so there are fewer question marks.  Some fields (user, group, permissions) would be blank anyway if the package had been successfully read, so make them blank for aborted packages.