Don't show values in the tree when stat has failed
If the read state is DirError or DirNoAccess, then the stat of that item failed and there is no meaningful data in most fields, so just leave them blank in the tree.  DirPermission denied items do have stat data, but any children will not and may not appear at all.  Sizes are already displayed with a prefix, although that will be "> 0 bytes" in many cases, and item counts are displayed as "?".