Create items in PkgReader even when stat() fails
Currently, if the call to fstatat() fails when PkgReader is adding files to the tree, no item is created.  The parent read state may be set to DirError or DirPermissionDenied even though there may be no problem with that directory.  Missing files are ignored completely, and only logged if a particular setting is true.
Instead, create DirInfo items when fstatat() returns a non-zero result, although with little beyond a name.  The package has those items in it, so the semi-abstract package tree should too.  Leave the parent alone and mark the new DirInfo read state as DirError, DirNoAccess, or DirMissing if it doesn't exist at all.
Display those in the file details panel with a type of "unknown" and a message of "[read error]", "[permission denied]", or "[missing]", and no directory node block.
UnreadableDirs picks up such items, but can't display anything about them, so it needs to be tweaked to list their parents instead.