Clean up stray trashinfo files
When a move to trash fails, for example for no permissions, the trashinfo file has already been created and remains.  Delete it.  The problem could be mostly solved by creating the trashinfo file last, but the spec says to create the trashinfo file with O_EXCL to make sure the name is unique.  So use QIODevice::NewOnly in QT 5.11 and newer instead of relying on other trash users to do it properly.  However, if the exclusive open fails due to a very unfortunate race condition (between checking if the name is unique and actually opening it), just fail instead of re-trying with a different name.
Also add formatErrno() to the most common failure message and reduce the logging verbosity a bit.  Most error messages are still logged three times, on on throw, once at caught, and once in the caller.