Simplify ProcessStarter
ProcessStarter doesn't currently start until it is explicitly told to, to prevent it deleting itself if the queue is starved.  It also has an auto-delete flag which controls whether it ever deletes itself.
Instead, start processes as soon as they are submitted, have an explicit notification that there will be no more processes,and always self-delete when the queue first becomes empty after that notification.  Fewer flags and the processes get started earlier.  This will be valuable for the new TrashWindow.