Presenting TrashWindow
A dedicated dialog and class that presents a merged list of items from all trash directories (currently mounted) for the current user.  The original item name and directory are shown, together with the total size (including directory children) and date deleted.  Items are only shown at the top level, so trashed directories don't show children.  Items lacking permissions or with other problems such as corrupt trashinfo are shown in red (configurable and sensitive to light/dark themes) with an error message.
Items in the dialog can be deleted (permanently) or restored (may fail, currently the only feedback is the log), with multi-selection allowed.
The dialog can be refreshed and the entire trash can be emptied (permanently!).