More consistent UI
Make the refresh shortcut in filesystems window alt-R for consistency.  That does mean that read is now alt-d, but it is also triggered by enter/return and item activation.
Fix the text of some context menu actions in file type stats window, for non-pattern items.  Also add a ctrl-R hotkey for refresh and put it on a context menu even when a valid item is not clicked.