Make nicer to use TrashWindow
Add actions and configurable hotkeys for TrashWindow, including "select all", and a context menu.
Add a message box when restoring an item to a name that already exists.  The only options are to replace it, or not.  Plus yes-to-all, no-to-all, and abort for dealing with multiple items.  The other likely error, missing parent directory, is addressed by creating it (and its ancestors), no questions, just do it.
Add an error popup for failed restores.
Disable restore action for items with errors, including multiple selections; the original path and name won't be known.  They can still be deleted, which is one of the main reasons for them being shown at all.