Buttons, focus, and defaults
Move the TrashWindow "Empty" button to top right and label it "Empty Trash" in bold.  Remove the alt-E shortcut, which is a little too easy to press unintentionally with no confirmation prompt.  The ctrl-shift-del hotkey remains, but it requires pressing three buttons and can be reconfigured or removed if desired.
Adjust the TrashWindow focus to always start in the tree, with the Refresh button becoming the (auto-)default although enter/return usually triggers the restore button through tree item activation or hotkey.
Tweak the auto button default or focus in some other window for a similar behaviour, with the default being refresh (close might be better?) and initial focus in the tree.