Un-optimise Qt6
QAbstractItemView::dataChanged has been optimised in Qt6 to only update the portion of the changed data rectangle that overlaps the visible viewport, instead of updating the whole viewport blindly in Qt5.  Unfortunately, during reads, the whole viewport needs to be updated or changes in the sort (by pending read jobs) order are not reflected.  The updated row itself may also not be updated, depending on whether the view managed to recognise the child rows before the tree was expanded to level 1.

The hacky fix is to update the whole viewport in an override of the dataChanged() function.  This is only done in Qt6.  Unfortunately, this carries the overhead of rectangle calculation and the overhead of re-painting the whole viewport for every update, but performance (ignoring i/o)  still seems slightly better than Qt6.  The alternative is either a fairly extensive re-design of the model update code or to abandon row-by-row updates completely in favour of layoutChanged.  Neither of these is appealing at the moment.