Better selection after TrashWindow operation
To select an item after deleting or restoring several items, using the index of the current item may result in an item being selected further down the list.  Instead, when multiple items are selected, use the index of the first, so that the selected item afterwards is at about the same position  Fall back to the current item when there is no selection.