TrashWindow cleanup
Improve the selection after a delete or restore of some items.  Instead of using the index of the current item, use the index of the first selected item.  This avoids the selection jumping down the list when the current item is not the first of multiple selected items.
Some re-factoring to reduce the size of the TrashItem constructor (a bit!) and make the validity check on a trashinfo file less frightening.
Add a dot to "trashinfo" in some error messages.