Consistency in "read error" fields shown
In the file details panel, the subtree fields are blank when there is a subtree message, except for the package summary mtime.  Make it blank also.  A better solution would be to show the subtree fields with a ">" prefix for consistency with the tree, but where to put the message is a problem.  The current layout makes the message stand out, and the missing information is dubious anyway.