Show allocated size in TrashWindow
Change the size shown in TrashWindow to be comparable to the ones shown in the main tree or in a normal du call (ie. not --apparent-size).  This is done by creating a temporary FileInfo object and using its allocatedSize() value.  The du call is modified to use "-B 1" instead of "-b".
Go back to use DirTreeModel to pick an icon since there is a FileInfo object available.