Make refresh update the line height
The line height in trees in various dialogs is configurable (small or medium).  The main tree responds to configuration changes in real time through DirTreeModel.  The various stats, etc. dialogs are not notified, but may get updated icons from DirTreeModel when they are refreshed.  Or not.

Make it consistent.  Always update the QAbstractItemView::iconSize (or PercentBarDelegate::height) when the tree is populated instead of just in initTree; that combined with the actual icons from DirTreeModel will consistently change the line height.  In some cases (eg. OpenDirDialog), there is no refresh and the tree line height will only change if the dialog is closed and re-opened.