Trash error handling
Provide a Trash member function to identify if a path lies within any Trash directory (not just the ones in the map).  CleanupCollection checks that the user has access to move a file to be trashed, then checks that it isn't within a trash directory which will fail (or create a mess) even if the permissions are apparently OK.

Quite a lot of re-factoring, for example moving trashRoots() from TrashWindow to Trash.  Also make TrashDir more self-contained, so that its main public function is simple trash() instead of two separate functions.

Also use QFile::errorString instead of formatErrno() in a couple of messages.